Move events and listening functions to libs
