+++ tests/monitoring/monitoring_service/test_end_to_end.py
@@ -2,11 +2,11 @@ import gevent
 from request_collector.server import RequestCollector
 from web3 import Web3
 
-from monitoring_service.blockchain import query_blockchain_events
 from monitoring_service.service import MonitoringService
 from raiden.utils.typing import BlockNumber
 from raiden_contracts.constants import CONTRACT_MONITORING_SERVICE, MonitoringServiceEvent
 from raiden_contracts.contract_manager import ContractManager
+from raiden_libs.blockchain import query_blockchain_events
 from raiden_libs.types import Address
